add math.pow and changes to use floats instead of int (not tested))

+++ src/Condition.cpp
@@ -29,8 +29,8 @@ Condition::Condition(
 };
     
 void Condition::execute() {
-    int left = leftIsValue ? game->values.find(leftID)->second.get() : game->calculations.find(leftID)->second.result();
-    int right = rightIsValue ? game->values.find(rightID)->second.get() : game->calculations.find(rightID)->second.result();
+    float left = leftIsValue ? game->values.find(leftID)->second.get() : game->calculations.find(leftID)->second.result();
+    float right = rightIsValue ? game->values.find(rightID)->second.get() : game->calculations.find(rightID)->second.result();
 
     bool validationResult = false;
     switch(conditionOperator) {
